Access product through variant explicitly
In Solidus 1.x there was a `delegate :product, to: :variant` on
Spree::LineItem. In 2.x this was removed, since there is already a
`has_one :product, through: :variant` on LineItem's.

This change causes issues in specs where we use the build strategy for
line_items because line_item.product fails because the variant_id on the
line_item is still nil. If we reference it explicitly, we can access the
product without first saving the record.